fix types + test refactoring
