+++ integrations/cohere/src/haystack_integrations/components/generators/cohere/chat/chat_generator.py
@@ -1,5 +1,5 @@
 import json
-from typing import Any, AsyncIterator, Dict, Iterator, List, Literal, Optional, TypedDict, Union, get_args
+from typing import Any, AsyncIterator, Dict, Iterator, List, Literal, Optional, Union, get_args
 
 from haystack import component, default_from_dict, default_to_dict, logging
 from haystack.components.generators.utils import _convert_streaming_chunks_to_chat_message
@@ -28,6 +28,8 @@ from cohere import (
     AsyncClientV2,
     ChatResponse,
     ClientV2,
+    ImageUrl,
+    ImageUrlContent,
     StreamedChatResponseV2,
     SystemChatMessageV2,
     TextAssistantMessageV2ContentItem,
@@ -51,32 +53,6 @@ ImageFormat = Literal["image/png", "image/jpeg", "image/webp", "image/gif"]
 IMAGE_SUPPORTED_FORMATS: list[ImageFormat] = list(get_args(ImageFormat))
 
 
-# TypedDict for multimodal content structure
-class ImageUrlDict(TypedDict):
-    url: str
-
-
-class ImageContentDict(TypedDict):
-    type: str
-    image_url: ImageUrlDict
-
-
-class TextContentDict(TypedDict):
-    type: str
-    text: str
-
-
-class MultimodalMessageDict(TypedDict):
-    role: str
-    content: List[Union[TextContentDict, ImageContentDict]]
-
-
-# Type alias for all possible message types
-FormattedMessage = Union[
-    UserChatMessageV2, AssistantChatMessageV2, SystemChatMessageV2, ToolChatMessageV2, MultimodalMessageDict
-]
-
-
 def _format_tool(tool: Tool) -> Dict[str, Any]:
     """
     Formats a Haystack Tool into Cohere's function specification format.
@@ -97,7 +73,9 @@ def _format_tool(tool: Tool) -> Dict[str, Any]:
     }
 
 
-def _format_message(message: ChatMessage) -> FormattedMessage:
+def _format_message(
+    message: ChatMessage,
+) -> Union[UserChatMessageV2, AssistantChatMessageV2, SystemChatMessageV2, ToolChatMessageV2]:
     """
     Formats a Haystack ChatMessage into Cohere's chat format.
 
@@ -108,12 +86,16 @@ def _format_message(message: ChatMessage) -> FormattedMessage:
     - Tool call results
 
     :param message: Haystack ChatMessage to format.
-    :return: A Cohere message object or MultimodalMessageDict for multimodal messages.
+    :return: A Cohere message object.
     """
     if not message.texts and not message.tool_calls and not message.tool_call_results and not message.images:
         msg = (
             "A `ChatMessage` must contain at least one `TextContent`, `ImageContent`, `ToolCall`, or `ToolCallResult`."
         )
+        raise ValueError(msg)
+
+    if message.images and not message.role.value == "user":
+        msg = "`ImageContent` is only supported for user messages."
         raise ValueError(msg)
 
     # Format the message based on its content type
@@ -123,7 +105,8 @@ def _format_message(message: ChatMessage) -> FormattedMessage:
             msg = "`ToolCall` must have a non-null `id` attribute to be used with Cohere."
             raise ValueError(msg)
         return ToolChatMessageV2(tool_call_id=result.origin.id, content=json.dumps({"result": result.result}))
-    elif message.tool_calls:
+
+    if message.tool_calls:
         tool_calls = []
         for tool_call in message.tool_calls:
             if tool_call.id is None:
@@ -143,57 +126,53 @@ def _format_message(message: ChatMessage) -> FormattedMessage:
             tool_calls=tool_calls,
             tool_plan=message.text if message.text else "",
         )
-    else:
-        if not message.texts or not message.texts[0]:
-            msg = (
-                "A `ChatMessage` must contain at least one `TextContent`, `ImageContent`, "
-                "`ToolCall`, or `ToolCallResult`."
-            )
+
+    if message.role.value == "user":
+        if not message.images and not message.text:
+            msg = "A `ChatMessage` from user must contain at least one non-empty `TextContent` or `ImageContent`."
             raise ValueError(msg)
 
-        # Check that images are only in user messages
-        if message.images and not message.is_from("user"):
-            error_msg = "Image content is only supported for user messages"
-            raise ValueError(error_msg)
+        # Handle multimodal content (text + images)
+        if message.images:
+            # Validate image formats
+            for image in message.images:
+                if image.mime_type not in IMAGE_SUPPORTED_FORMATS:
+                    supported_formats = ", ".join(IMAGE_SUPPORTED_FORMATS)
+                    msg = (
+                        f"Unsupported image format: {image.mime_type}. "
+                        f"Cohere supports the following formats: {supported_formats}"
+                    )
+                    raise ValueError(msg)
 
-        if message.role.value == "user":
-            # Handle multimodal content (text + images)
-            if message.images:
-                # Validate image formats
-                for image in message.images:
-                    if image.mime_type not in IMAGE_SUPPORTED_FORMATS:
-                        supported_formats = ", ".join(IMAGE_SUPPORTED_FORMATS)
-                        msg = (
-                            f"Unsupported image format: {image.mime_type}. "
-                            f"Cohere supports the following formats: {supported_formats}"
-                        )
-                        raise ValueError(msg)
+        # Build multimodal content following Cohere's API specification
+        content_parts: List[Union[CohereTextContent, ImageUrlContent]] = []
+        for part in message._content:
+            if isinstance(part, TextContent) and part.text:
+                text_content = CohereTextContent(text=part.text)
+                content_parts.append(text_content)
+            elif isinstance(part, ImageContent):
+                # Cohere expects base64 data URI format
+                # See: https://docs.cohere.com/docs/image-inputs
+                image_url = f"data:{part.mime_type};base64,{part.base64_image}"
+                image_content = ImageUrlContent(image_url=ImageUrl(url=image_url))
+                content_parts.append(image_content)
 
-                # Build multimodal content following Cohere's API specification
-                content_parts: List[Union[TextContentDict, ImageContentDict]] = []
-                for part in message._content:
-                    if isinstance(part, TextContent) and part.text:
-                        text_content: TextContentDict = {"type": "text", "text": part.text}
-                        content_parts.append(text_content)
-                    elif isinstance(part, ImageContent):
-                        # Cohere expects base64 data URI format
-                        # See: https://docs.cohere.com/docs/image-inputs
-                        image_url = f"data:{part.mime_type};base64,{part.base64_image}"
-                        image_content: ImageContentDict = {"type": "image_url", "image_url": {"url": image_url}}
-                        content_parts.append(image_content)
+        return UserChatMessageV2(content=content_parts)
 
-                # Return typed dict for multimodal messages
-                multimodal_message: MultimodalMessageDict = {"role": "user", "content": content_parts}
-                return multimodal_message
-            else:
-                return UserChatMessageV2(content=[CohereTextContent(text=message.texts[0])])
-        elif message.role.value == "assistant":
-            return AssistantChatMessageV2(content=[TextAssistantMessageV2ContentItem(text=message.texts[0])])
-        elif message.role.value == "system":
-            return SystemChatMessageV2(content=[TextSystemMessageV2ContentItem(text=message.texts[0])])
-        else:
-            msg = f"Unsupported message role: {message.role.value}"
+    if message.role.value == "assistant":
+        if not message.text:
+            msg = "A `ChatMessage` from assistant without tool calls must contain a non-empty `TextContent`."
             raise ValueError(msg)
+        return AssistantChatMessageV2(content=[TextAssistantMessageV2ContentItem(text=message.text)])
+
+    if message.role.value == "system":
+        if not message.text:
+            msg = "A `ChatMessage` from system calls must contain a non-empty `TextContent`."
+            raise ValueError(msg)
+        return SystemChatMessageV2(content=[TextSystemMessageV2ContentItem(text=message.text)])
+
+    msg = f"Unsupported message role: {message.role.value}"
+    raise ValueError(msg)
 
 
 def _parse_response(chat_response: ChatResponse, model: str) -> ChatMessage:
@@ -285,7 +264,7 @@ def _convert_cohere_chunk_to_streaming_chunk(
     start = False
     finish_reason = None
     tool_calls = None
-    meta = {"model": model}
+    meta: Dict[str, Any] = {"model": model}
 
     if chunk.type == "content-delta" and chunk.delta and chunk.delta.message:
         if chunk.delta.message and chunk.delta.message.content and chunk.delta.message.content.text is not None:
@@ -309,7 +288,8 @@ def _convert_cohere_chunk_to_streaming_chunk(
                     )
                 ]
                 start = True  # This starts a tool call
-                meta["tool_call_id"] = tool_call.id  # type: ignore[assignment]
+                if tool_call.id is not None:
+                    meta["tool_call_id"] = tool_call.id
 
     elif chunk.type == "tool-call-delta" and chunk.delta and chunk.delta.message:
         if (
@@ -356,8 +336,8 @@ def _convert_cohere_chunk_to_streaming_chunk(
 
         usage = {"prompt_tokens": prompt_tokens, "completion_tokens": completion_tokens}
 
-        meta["finish_reason"] = finish_reason_raw  # type: ignore[assignment]
-        meta["usage"] = usage  # type: ignore[assignment]
+        meta["finish_reason"] = finish_reason_raw
+        meta["usage"] = usage
 
     return StreamingChunk(
         content=content,
@@ -670,7 +650,7 @@ class CohereChatGenerator:
             _check_duplicate_tool_names(tools)
             generation_kwargs["tools"] = [_format_tool(tool) for tool in tools]
 
-        formatted_messages: List[FormattedMessage] = [_format_message(message) for message in messages]
+        formatted_messages = [_format_message(message) for message in messages]
 
         streaming_callback = select_streaming_callback(
             init_callback=self.streaming_callback, runtime_callback=streaming_callback, requires_async=False
@@ -680,7 +660,7 @@ class CohereChatGenerator:
             component_info = ComponentInfo.from_component(self)
             streamed_response = self.client.chat_stream(
                 model=self.model,
-                messages=formatted_messages,  # type: ignore[arg-type]  # Client accepts MultimodalMessageDict at runtime
+                messages=formatted_messages,
                 **generation_kwargs,
             )
             chat_message = _parse_streaming_response(
@@ -692,7 +672,7 @@ class CohereChatGenerator:
         else:
             response = self.client.chat(
                 model=self.model,
-                messages=formatted_messages,  # type: ignore[arg-type]  # Client accepts MultimodalMessageDict at runtime
+                messages=formatted_messages,
                 **generation_kwargs,
             )
             chat_message = _parse_response(response, self.model)
@@ -736,7 +716,7 @@ class CohereChatGenerator:
             _check_duplicate_tool_names(tools)
             generation_kwargs["tools"] = [_format_tool(tool) for tool in tools]
 
-        formatted_messages: List[FormattedMessage] = [_format_message(message) for message in messages]
+        formatted_messages = [_format_message(message) for message in messages]
 
         streaming_callback = select_streaming_callback(
             init_callback=self.streaming_callback, runtime_callback=streaming_callback, requires_async=True
@@ -746,7 +726,7 @@ class CohereChatGenerator:
             component_info = ComponentInfo.from_component(self)
             streamed_response = self.async_client.chat_stream(
                 model=self.model,
-                messages=formatted_messages,  # type: ignore[arg-type]  # Client accepts MultimodalMessageDict at runtime
+                messages=formatted_messages,
                 **generation_kwargs,
             )
             chat_message = await _parse_async_streaming_response(
@@ -758,7 +738,7 @@ class CohereChatGenerator:
         else:
             response = await self.async_client.chat(
                 model=self.model,
-                messages=formatted_messages,  # type: ignore[arg-type]  # Client accepts MultimodalMessageDict at runtime
+                messages=formatted_messages,
                 **generation_kwargs,
             )
             chat_message = _parse_response(response, self.model)
